updated crud methods.
decrease loan when deposit mafe, customer post only cust details, post loan using cust id, need to fix xml post requests

+++ Loans/src/main/java/crudService/DepositService.java
@@ -5,13 +5,17 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import dao.DepositDAO;
+import dao.LoanDAO;
+import entities.Customer;
 import entities.Deposit;
+import entities.Loan;
 
 import java.util.List;
 
 @Path("/deposits")
 public class DepositService {
     private DepositDAO dao = new DepositDAO();
+    private LoanDAO loanDao = new LoanDAO();
 
     @GET
     @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
@@ -30,11 +34,24 @@ public class DepositService {
             return Response.status(Response.Status.NOT_FOUND).build();
     }
 
+    
     @POST
+    @Path("/{loanId}")
     @Consumes({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
     @Produces({MediaType.APPLICATION_JSON, MediaType.APPLICATION_XML})
-    public Response addDeposit(Deposit deposit) {
-        dao.persist(deposit);
+    public Response addDeposit(@PathParam("loanId") int loanId, Deposit deposit) {
+        Loan loan = loanDao.getLoanById(loanId);
+        if (loan != null) {
+            double newLoanAmount = loan.getLoanAmount() - deposit.getAmount();
+            if (newLoanAmount < 0) newLoanAmount = 0;
+            loan.setLoanAmount(newLoanAmount);
+          
+            loan.getDeposits().add(deposit);
+            
+            loanDao.merge(loan);
+        } else {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
         return Response.ok(deposit).build();
     }
 
@@ -57,10 +74,21 @@ public class DepositService {
     public Response deleteDeposit(@PathParam("id") int id) {
         Deposit deposit = dao.getDepositById(id);
         if (deposit != null) {
-            dao.remove(deposit);
+           
+            Loan loan = loanDao.getLoanByDeposit(deposit); 
+            if (loan != null) {
+                loan.getDeposits().remove(deposit);
+                double newLoanAmount = loan.getLoanAmount() + deposit.getAmount();
+                loan.setLoanAmount(newLoanAmount);
+                loanDao.merge(loan);
+                dao.remove(deposit);
+            } else {
+                return Response.status(Response.Status.NOT_FOUND).entity("loan not found").build();
+            }
+
             return Response.ok().build();
         } else {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            return Response.status(Response.Status.NOT_FOUND).entity("Deposit not found").build();
         }
     }
 }
